+++ src/hr/avrbanac/openglplayground/normalMappingRenderer/NormalMappingRenderer.java
@@ -1,0 +1,100 @@
+
+package hr.avrbanac.openglplayground.normalMappingRenderer;
+
+import static hr.avrbanac.openglplayground.Globals.*;
+import hr.avrbanac.openglplayground.entities.Camera;
+import hr.avrbanac.openglplayground.entities.Entity;
+import hr.avrbanac.openglplayground.entities.Light;
+import hr.avrbanac.openglplayground.maths.Matrix4f;
+import hr.avrbanac.openglplayground.maths.Vector4f;
+import hr.avrbanac.openglplayground.models.RawModel;
+import hr.avrbanac.openglplayground.models.TexturedModel;
+import hr.avrbanac.openglplayground.renderers.MasterRenderer;
+import hr.avrbanac.openglplayground.textures.ModelTexture;
+import java.util.List;
+import java.util.Map;
+import org.lwjgl.opengl.GL11;
+import org.lwjgl.opengl.GL13;
+import org.lwjgl.opengl.GL20;
+import org.lwjgl.opengl.GL30;
+
+/**
+ *
+ * @author avrbanac
+ * @version 1.0.18
+ */
+public class NormalMappingRenderer {
+    private final NormalMappingShader shader;
+ 
+    public NormalMappingRenderer(Matrix4f projectionMatrix) {
+        this.shader = new NormalMappingShader();
+        shader.start();
+        shader.loadProjectionMatrix(projectionMatrix);
+        shader.connectTextureUnits();
+        shader.stop();
+    }
+ 
+    public void render(Map<TexturedModel, List<Entity>> entities, Vector4f clipPlane, List<Light> lights, Camera camera) {
+        shader.start();
+        prepare(clipPlane, lights, camera);
+        for (TexturedModel model : entities.keySet()) {
+            prepareTexturedModel(model);
+            List<Entity> batch = entities.get(model);
+            for (Entity entity : batch) {
+                prepareInstance(entity);
+                GL11.glDrawElements(GL11.GL_TRIANGLES, model.getRawModel().getVertexCount(), GL11.GL_UNSIGNED_INT, 0);
+            }
+            unbindTexturedModel();
+        }
+        shader.stop();
+    }
+     
+    public void cleanUp(){
+        shader.cleanUp();
+    }
+ 
+    private void prepareTexturedModel(TexturedModel model) {
+        RawModel rawModel = model.getRawModel();
+        GL30.glBindVertexArray(rawModel.getVaoID());
+        GL20.glEnableVertexAttribArray(0);
+        GL20.glEnableVertexAttribArray(1);
+        GL20.glEnableVertexAttribArray(2);
+        GL20.glEnableVertexAttribArray(3);
+        ModelTexture texture = model.getTexture();
+        shader.loadNumberOfRows(texture.getNumberOfRows());
+        if (texture.isTransparency()) {
+            MasterRenderer.disableCulling();
+        }
+        shader.loadShineVariables(texture.getShineDamper(), texture.getReflectivity());
+        GL13.glActiveTexture(GL13.GL_TEXTURE0);
+        GL11.glBindTexture(GL11.GL_TEXTURE_2D, model.getTexture().getTextureID());
+        GL13.glActiveTexture(GL13.GL_TEXTURE1);
+        GL11.glBindTexture(GL11.GL_TEXTURE_2D, model.getTexture().getNormalMap());
+    }
+ 
+    private void unbindTexturedModel() {
+        MasterRenderer.enableCulling();
+        GL20.glDisableVertexAttribArray(0);
+        GL20.glDisableVertexAttribArray(1);
+        GL20.glDisableVertexAttribArray(2);
+        GL20.glDisableVertexAttribArray(3);
+        GL30.glBindVertexArray(0);
+    }
+ 
+    private void prepareInstance(Entity entity) {
+        Matrix4f transformationMatrix = Matrix4f.transformation(entity.getPosition(), entity.getRotX(),
+                entity.getRotY(), entity.getRotZ(), entity.getScale());
+        shader.loadTransformationMatrix(transformationMatrix);
+        shader.loadOffset(entity.getTextureXOffset(), entity.getTextureYOffset());
+    }
+ 
+    private void prepare(Vector4f clipPlane, List<Light> lights, Camera camera) {
+        shader.loadClipPlane(clipPlane);
+        //need to be public variables in MasterRenderer
+        shader.loadSkyColour(SKY_RED, SKY_GREEN, SKY_BLUE);
+        Matrix4f viewMatrix = Matrix4f.view(camera);
+         
+        shader.loadLights(lights, viewMatrix);
+        shader.loadViewMatrix(viewMatrix);
+    }
+}